docs/src/requirements.txt
absl-py>=0.5.0
numpy>=1.13.3
tensorflow>=1.13.0
tensorflow-serving-api>=1.13.0
gin-config>=0.1.4
tensorflow-probability>=0.6.0
tf-agents>=0.3.0
tf-slim>=1.0
git+https://github.com/google-research/tensor2robot#tensor2robot
